Adding add2home css and javascript [Completes #38680093]
Note: this also modifies the behavior of the asset pipeline slightly.

Before all CSS and all JS were loaded on all page loads. Now, everything in application is loaded on all page loads.

Controller specific assets are loaded on all views for that controller.

Action specific assets are loaded for all views for that aciton.

E.g., application.js, home.js, home/index.js are all loaded on /.

This allows beta signup logic to only be loaded on the front page, and add2home to only be loaded on the dashboard.

cc @dannychapman