Adding extra logic for additional or missing data returned from API
I added county level representatives for the whole country, but since
that data is missing common things like URLs I had to be more explicit
about when those would be available. I'm also filtering out the County
level reps here because the only information available for them is
their name and for each county it's 5-10 people, so I'm currently
opting to just leave those out of this view for now.